+++ src/components/WeekTime/WeekTimeSelector/WeekTimeSelector.tsx
@@ -1,10 +1,14 @@
 import * as Styled from "./styled";
-import ErrorText from "../../ErrorText";
-import React, { useEffect, useMemo, useState } from "react";
-import { convertTargetSettingToHex, getNewWeekTimeList } from "./utils";
+import React, { Fragment, useEffect, useMemo, useState } from "react";
+import {
+  convertTargetSettingToHex,
+  getNewWeekTimeList,
+} from "./utils";
+import { useLocaleProps } from "../../../hooks/useLocaleProps";
 import { getTargetSetting } from "../utils";
+import { timeList } from "../constants";
 import Spacer from "../../Spacer";
-import { WeekTimeElement } from "../internal/WeekTimeElement";
+import ErrorText from "../../ErrorText";
 
 export type WeekTimeSelectorProps = {
   weekTime: string;
@@ -13,112 +17,129 @@ export type WeekTimeSelectorProps = {
 };
 
 const WeekTimeSelector = React.forwardRef<HTMLDivElement, WeekTimeSelectorProps>(
-  function WeekTimeSelector(props, ref) {
-    const { weekTime, errorText, onChange } = props;
-
-  const weekTimeList = useMemo(() => getTargetSetting(weekTime), [weekTime]);
-  const [startIndex, setStartIndex] = useState<{
-    weekIndex: number;
-    timeIndex: number;
-  }>({ weekIndex: 0, timeIndex: 0 });
-  const [endIndex, setEndIndex] = useState<{
-    weekIndex: number;
-    timeIndex: number;
-  }>({ weekIndex: 0, timeIndex: 0 });
-  const [selectState, setSelectState] = useState<"none" | "choosing">("none");
-  const [toValue, setToValue] = useState<string>("1");
-
-  const handleChangeWeekTime = (
-    weekIndex: number,
-    timeIndex: number,
-    newValue: string,
-  ) => {
-    if (selectState === "choosing") {
-      const newWeekTimeList = getNewWeekTimeList(
-        startIndex.weekIndex,
-        startIndex.timeIndex,
-        weekIndex,
-        timeIndex,
-        newValue,
-        [...weekTimeList],
-      );
-
-      setSelectState("none");
-
-      onChange?.(convertTargetSettingToHex(newWeekTimeList));
-    } else if (selectState === "none") {
-      setStartIndex({ weekIndex, timeIndex });
-      setEndIndex({ weekIndex, timeIndex });
-      setSelectState("choosing");
-    }
-  };
-
-  const isWithinHoverRange = (weekIndex: number, timeIndex: number) => {
-    if (selectState === "choosing") {
-      const minWeekIndex = Math.min(startIndex.weekIndex, endIndex.weekIndex);
-      const maxWeekIndex = Math.max(startIndex.weekIndex, endIndex.weekIndex);
-      const minTimeIndex = Math.min(startIndex.timeIndex, endIndex.timeIndex);
-      const maxTimeIndex = Math.max(startIndex.timeIndex, endIndex.timeIndex);
-
-      return (
-        weekIndex >= minWeekIndex &&
-        weekIndex <= maxWeekIndex &&
-        timeIndex >= minTimeIndex &&
-        timeIndex <= maxTimeIndex
-      );
-    }
-    return false;
-  };
-
-  const handleMouseOver = (weekIndex: number, timeIndex: number) => {
-    if (selectState === "none") {
-      return;
-    }
-
-    setEndIndex({ weekIndex, timeIndex });
-  };
-
-  const handleMouseDown = (
-    weekIndex: number,
-    timeIndex: number,
-    time: string,
-  ) => {
-    const newValue = time === "1" ? "0" : "1";
-    setToValue(newValue);
-    handleChangeWeekTime(weekIndex, timeIndex, newValue);
-  };
-
-  // MEMO: button 以外の場所で mouseup したときの挙動を制御する
-  const onMouseUp = () => {
-    // mousedown の位置が button のときにのみこの条件に到達する想定
-    if (selectState === "choosing") {
-      handleChangeWeekTime(endIndex.weekIndex, endIndex.timeIndex, toValue);
-    }
-  };
-
-  useEffect(() => {
-    addEventListener("mouseup", onMouseUp);
-    return () => {
-      removeEventListener("mouseup", onMouseUp);
+  function WeekTimeSelector(inProps, ref) {
+    const props = useLocaleProps({
+      props: inProps,
+      name: "WeekTimeSelector",
+    });
+  
+    const { weekList, weekTime, errorText, onChange } = props;
+  
+    const weekTimeList = useMemo(() => getTargetSetting(weekTime), [weekTime]);
+    const [startIndex, setStartIndex] = useState<{
+      weekIndex: number;
+      timeIndex: number;
+    }>({ weekIndex: 0, timeIndex: 0 });
+    const [endIndex, setEndIndex] = useState<{
+      weekIndex: number;
+      timeIndex: number;
+    }>({ weekIndex: 0, timeIndex: 0 });
+    const [selectState, setSelectState] = useState<"none" | "choosing">("none");
+    const [toValue, setToValue] = useState<string>("1");
+  
+    const handleChangeWeekTime = (
+      weekIndex: number,
+      timeIndex: number,
+      newValue: string,
+    ) => {
+      if (selectState === "choosing") {
+        const newWeekTimeList = getNewWeekTimeList(
+          startIndex.weekIndex,
+          startIndex.timeIndex,
+          weekIndex,
+          timeIndex,
+          newValue,
+          [...weekTimeList],
+        );
+  
+        setSelectState("none");
+  
+        onChange?.(convertTargetSettingToHex(newWeekTimeList));
+      } else if (selectState === "none") {
+        setStartIndex({ weekIndex, timeIndex });
+        setEndIndex({ weekIndex, timeIndex });
+        setSelectState("choosing");
+      }
     };
-  });
-
-  return (
-    <Styled.Container>
-      <WeekTimeElement
-        weekTime={weekTime}
-        isWithinHoverRange={isWithinHoverRange}
-        onMouseOver={handleMouseOver}
-        onMouseDown={handleMouseDown}
-        ref={ref}
-      />
-      {errorText && (
-        <Spacer pt={1}>
-          <ErrorText>{errorText}</ErrorText>
-        </Spacer>
-      )}
-    </Styled.Container>
-  );
+  
+    const isWithinHoverRange = (weekIndex: number, timeIndex: number) => {
+      if (selectState === "choosing") {
+        const minWeekIndex = Math.min(startIndex.weekIndex, endIndex.weekIndex);
+        const maxWeekIndex = Math.max(startIndex.weekIndex, endIndex.weekIndex);
+        const minTimeIndex = Math.min(startIndex.timeIndex, endIndex.timeIndex);
+        const maxTimeIndex = Math.max(startIndex.timeIndex, endIndex.timeIndex);
+  
+        return (
+          weekIndex >= minWeekIndex &&
+          weekIndex <= maxWeekIndex &&
+          timeIndex >= minTimeIndex &&
+          timeIndex <= maxTimeIndex
+        );
+      }
+      return false;
+    };
+  
+    const handleMouseOver = (weekIndex: number, timeIndex: number) => () => {
+      if (selectState === "none") {
+        return;
+      }
+  
+      setEndIndex({ weekIndex, timeIndex });
+    };
+  
+    const handleMouseDown =
+      (weekIndex: number, timeIndex: number, time: string) => () => {
+        const newValue = time === "1" ? "0" : "1";
+        setToValue(newValue);
+        handleChangeWeekTime(weekIndex, timeIndex, newValue);
+      };
+  
+    // MEMO: button 以外の場所で mouseup したときの挙動を制御する
+    const onMouseUp = () => {
+      // mousedown の位置が button のときにのみこの条件に到達する想定
+      if (selectState === "choosing") {
+        handleChangeWeekTime(endIndex.weekIndex, endIndex.timeIndex, toValue);
+      }
+    };
+  
+    useEffect(() => {
+      addEventListener("mouseup", onMouseUp);
+      return () => {
+        removeEventListener("mouseup", onMouseUp);
+      };
+    });
+  
+    return (
+      <Styled.Container ref={ref}>
+        <Styled.WeekTimeContainer>
+          <Styled.EmptyContainer />
+          {timeList.map((time) => (
+            <Styled.TimeContainer key={time}>{time}</Styled.TimeContainer>
+          ))}
+          {weekTimeList.map((time, weekIndex) => (
+            // eslint-disable-next-line react/no-array-index-key
+            <Fragment key={weekIndex}>
+              <Styled.WeekContainer>{weekList[weekIndex]}</Styled.WeekContainer>
+              {time.map((t, timeIndex) => (
+                <Styled.WeekTimeItem
+                  // eslint-disable-next-line react/no-array-index-key
+                  key={`${weekIndex}-${timeIndex}`}
+                  active={t === "1"}
+                  hover={isWithinHoverRange(weekIndex, timeIndex)}
+                  onMouseOver={handleMouseOver(weekIndex, timeIndex)}
+                  onMouseDown={handleMouseDown(weekIndex, timeIndex, t)}
+                />
+              ))}
+            </Fragment>
+          ))}
+        </Styled.WeekTimeContainer>
+        {errorText && (
+          <Spacer pt={1}>
+            <ErrorText>{errorText}</ErrorText>
+          </Spacer>
+        )}
+      </Styled.Container>
+    );
   }
 )
