--- src/components/WeekTime/internal/WeekTimeElement/index.ts
@@ -1,1 +1,0 @@
-export { WeekTimeElement } from "./WeekTimeElement";
